agentcore/requirements.txt
boto3>=1.42.44,<2.0.0
bedrock-agentcore>=1.2.1,<2.0.0
strands-agents[otel]>=1.25.0,<2.0.0
langfuse>=3.7.0,<4.0.0
